implemented selector and got app working again with redux
