+++ src/containers/Moods.js
@@ -1,31 +1,16 @@
-import React, { Component } from 'react';
-import { useDispatch } from 'react-redux';
+import React from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import Controls from '../components/controls/Controls';
 import Face from '../components/face/Face';
 import { takeNap, drinkCoffee, study, eatSnack } from '../actions/moodActions';
-
-export const isTired = state => state.coffees < 1 && state.naps < 1;
-export const isHyper = state => state.coffees > 3;
-export const isEducated = state => state.studies > 2;
-export const isHungry = state => state.snacks < 1;
-
-export const getFace = state => {
-  if(isTired(state) && isHungry(state)) return '🤬';
-  if(isHyper(state) && isHungry(state)) return '🤮';
-  if(isTired(state)) return '😴';
-  if(isHyper(state)) return '🙀';
-  if(isEducated(state)) return '🤯';
-  if(isHungry(state)) return '😡';
-
-  return '😀';
-};
+import { getMood, getFace } from '../selectors/moodSelectors';
 
 export const Moods = () => {
 
   const dispatch = useDispatch();
-  
-  const { coffees, snacks, naps, studies } = this.state;
-  const face = getFace(this.state);
+
+  const { coffees, snacks, naps, studies } = useSelector(getMood);
+  const face = useSelector(getFace);
 
   return (
     <>
