+++ src/store.js
@@ -1,7 +1,7 @@
 import { createStore } from 'redux';
 import { moodReducer } from './reducers/moodReducer';
 
-export default createStore(
+export const store = createStore(
   moodReducer,
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
